Lab1/gateway/requirements.txt
asttokens==2.4.0
blinker==1.6.2
cachetools==5.3.1
circuitbreaker==2.0.0
click==8.1.7
colorama==0.4.6
executing==2.0.0
Flask==3.0.0
Flask-Executor==1.0.0
grpcio==1.59.0
grpcio-tools==1.59.0
icecream==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
protobuf==4.24.3
Pygments==2.16.1
pymemcache==4.0.0
python-dotenv==1.0.0
six==1.16.0
uhashring==2.3
Werkzeug==3.0.0